map update : add data stats

+++ app/src/main/java/com/example/mindtrip/MainActivity.kt
@@ -1,11 +1,14 @@
 package com.example.mindtrip
 
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.Intent
+import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
-import androidx.core.view.isVisible
+import android.widget.Button
+import android.widget.TableRow
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.mindtrip.databinding.ActHomeBinding
 import java.io.File
@@ -21,19 +24,7 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = ActHomeBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        if(initData()){
-            binding.clearbtn.visibility = View.VISIBLE
-            binding.recyclerView.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
-            adapter = ActAdapter(actdata)
-            adapter.itemClickListener = object : ActAdapter.OnItemClickListener {
-                override fun OnItemClick(data: ActData) {
-//                val intent = Intent(this@MainActivity, MapActivity::class.java)
-//                startActivity(intent)
-
-                }
-            }
-            binding.recyclerView.adapter = adapter
-        }
+        initRecyclerView()
         init()
 
     }
@@ -47,6 +38,50 @@ class MainActivity : AppCompatActivity() {
             return false
         }
     }
+
+    private fun initRecyclerView(){
+        actdata.clear()
+        binding.recyclerView.adapter?.notifyDataSetChanged()
+        if(initData()){
+            binding.clearbtn.visibility = View.VISIBLE
+            binding.refreshbtn.visibility = View.VISIBLE
+            binding.refreshbtn2.visibility = View.GONE
+            binding.recyclerView.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
+            adapter = ActAdapter(actdata)
+            adapter.itemClickListener = object : ActAdapter.OnItemClickListener {
+                override fun OnItemClick(data: ActData) {
+
+                    val builder = AlertDialog.Builder(this@MainActivity)
+                    builder!!.setMessage("Google Maps에 볼까요?")
+                        .setTitle(data.name)
+
+                    builder.apply {
+                        setPositiveButton("예") { dialog, id ->
+                            val selectedId = id
+                            val gmmIntentUri = Uri.parse("geo:0,0?q=${data.name}")
+                            val mapIntent = Intent(Intent.ACTION_VIEW, gmmIntentUri)
+                            startActivity(mapIntent)
+
+                        }
+                        setNegativeButton("아니요") { dialog, id ->
+                            val selectedId = id
+
+                        }
+                    }
+                    val dialog: AlertDialog? = builder.create()
+
+                    dialog!!.show()
+
+
+                }
+            }
+            binding.recyclerView.adapter = adapter
+        }else{
+            binding.clearbtn.visibility = View.GONE
+            binding.refreshbtn.visibility = View.GONE
+            binding.refreshbtn2.visibility = View.VISIBLE
+        }
+    }
     private fun init() {
 
         with(binding) {
@@ -55,12 +90,38 @@ class MainActivity : AppCompatActivity() {
                 startActivity(intent)
             }
             clearbtn.setOnClickListener {
-                actdata.clear()
-                binding.recyclerView.adapter?.notifyDataSetChanged()
-                val file = File("/data/data/com.example.mindtrip/files/statistics.txt")
-                file.delete()
-                clearbtn.visibility = View.GONE
+                val builder = AlertDialog.Builder(this@MainActivity)
+                builder!!.setMessage("활동 데이터 채울까요?")
+                    .setTitle("채우기")
 
+                builder.apply {
+                    setPositiveButton("예") { dialog, id ->
+                        val selectedId = id
+                        actdata.clear()
+                        binding.recyclerView.adapter?.notifyDataSetChanged()
+                        val file = File("/data/data/com.example.mindtrip/files/statistics.txt")
+                        file.delete()
+                        clearbtn.visibility = View.GONE
+                        refreshbtn.visibility = View.GONE
+                        refreshbtn2.visibility = View.VISIBLE
+
+                    }
+                    setNegativeButton("아니요") { dialog, id ->
+                        val selectedId = id
+
+                    }
+                }
+                val dialog: AlertDialog? = builder.create()
+
+                dialog!!.show()
+
+
+            }
+            refreshbtn2.setOnClickListener {
+                initRecyclerView()
+            }
+            refreshbtn.setOnClickListener {
+                initRecyclerView()
             }
         }
     }
